send tokens to commands instead of strings
Change `ParserCommand` to take `Token`s instead of strings. The tokens
store information on the location of the strings within a file, and so
can provide better debugging information, as well as being a step
towards better tooling.

For now, except for a few commands which forward their commands to
be further parsed and executed, strip the strings out of the tokens
and use them like before. This is just to reduce the amount of work
that must be done immediately to keep a working system. Each command
should eventually be changed to use the tokens properly.

Another design possibility was investigated and abandoned, which was
changing `ParserCommand` to allow the parsing of either tokens
or strings and not changing the majority of the `ParserCommand`
implementations. This was unsatisfactory for at least two reasons.
One, requiring ParserCommand implementations to implement one or the
other of the Parse commands was confusing, and changed compile-time
obvious checking to run-time not-so-obvious checking. Second, it
required strings and tokens to be converted back and forth, which was
both confusing and less efficient than the one-time optional
conversion which is used in the chosen design.